Restructure code
CMake now also GLOBs. This isn't a bad thing, as we're using CONFIGURE_DEPENDS. 3DStris already depends on CMake 3.12, so this change shouldn't break anything.
This is also a bit of an include cleanup as well.
