+++ include/3dstris/config/l10n.hpp
