+++ include/3dstris/gui.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <3dstris/gui/button.hpp>
-#include <3dstris/gui/widget.hpp>
 #include <3dstris/util.hpp>
 #include <vector>
