+++ include/3dstris/states/menu/loadfailed.hpp
@@ -2,7 +2,7 @@
 
 #include <3dstris/gui.hpp>
 #include <3dstris/state.hpp>
-#include <3dstris/states/playing.hpp>
+#include <3dstris/states/game/playing.hpp>
 #include <3dstris/util.hpp>
 
 class LoadFailed : public State {
