+++ include/3dstris/util/colorstextures.hpp
