+++ include/3dstris/util/text.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
 #include <sds.h>
-#include <3dstris/colors.hpp>
-#include <3dstris/util.hpp>
+#include <3dstris/util/colorstextures.hpp>
 
 class Text {
    public:
