+++ src/game.cpp
@@ -1,7 +1,7 @@
 #include <3dstris/images.h>
-#include <3dstris/colors.hpp>
 #include <3dstris/game.hpp>
 #include <3dstris/state.hpp>
+#include <3dstris/util/colorstextures.hpp>
 
 Game::Game()
 	: imageSheet(C2D_SpriteSheetLoad("romfs:/gfx/images.t3x")),
