+++ src/states/menu/languageselect.cpp
@@ -1,6 +1,6 @@
 #include <3dstris/gui/languagebutton.hpp>
-#include <3dstris/states/languages.hpp>
-#include <3dstris/states/mainmenu.hpp>
+#include <3dstris/states/menu/languageselect.hpp>
+#include <3dstris/states/menu/mainmenu.hpp>
 
 LanguageSelect::LanguageSelect()
 	: State(),
