+++ src/states/menu/sprinttimes.cpp
@@ -1,5 +1,5 @@
 #include <3dstris/game.hpp>
-#include <3dstris/states/sprinttimes.hpp>
+#include <3dstris/states/menu/sprinttimes.hpp>
 
 SprintTimes::SprintTimes()
 	: State(),
